+++ projects/ui-storybook/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @cmpsr/ui-storybook
 
+## 0.0.7
+
+### Patch Changes
+
+- Updated dependencies [f292607]
+  - @cmpsr/components@3.0.0
+
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [adba0b1]
+  - @cmpsr/components@2.4.0
+
 ## 0.0.5
 
 ### Patch Changes